requirements.txt
certifi==2018.8.24
chardet==3.0.4
Django==2.1
djangorestframework==3.8.2
dropbox==9.0.0
gunicorn==19.9.0
idna==2.7
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
python-decouple==3.1
pytz==2018.5
requests==2.19.1
six==1.11.0
urllib3==1.23
virtualenv==16.0.0
